feat(nav): Retain original menu attributes
Previous iterations of the library used the original menu as a guide to
construct a clone. This resulted in any existing classes or attributes
not being carried over.

This commit now clones the original menu directly, enhancing it with our
desired classes/data-attributes. Event listeners are not preserved due
to the cloning.